feat: move discount component to cartTable
