+++ src/components/CartSection/CartSection.module.scss
@@ -1,15 +1,30 @@
 @use '@/styles/variables' as vars;
 @use '@/styles/mixins' as mixins;
 
+.section {
+  display: flex;
+  flex-direction: row;
+  gap: 1rem;
+  align-items: flex-start;
+  justify-content: space-between;
+
+  @media (width <= 700px) {
+    flex-direction: column;
+    align-items: end;
+  }
+}
+
 .image {
   display: inline-block;
-  width: 6rem;
-  height: 6rem;
+  width: 3rem;
+  height: 3rem;
 
   & img {
-    width: 100%;
-    height: 100%;
     object-fit: contain;
+  }
+
+  @media (width <= 530px) {
+    display: none;
   }
 }
 
@@ -34,4 +49,12 @@
   align-items: center;
   justify-content: center;
   padding: 1rem 0 0;
+}
+
+.total {
+  display: block;
+  min-width: 12rem;
+  padding: 1rem;
+  text-align: center;
+  background-color: vars.$color-primary25;
 }
