+++ src/styles/base.scss
@@ -57,7 +57,7 @@ td {
   padding: 0.5rem;
   text-align: center;
 
-  & > div {
+  & div {
     display: flex;
     flex-direction: row;
     gap: 0.5rem;
